+++ nexusvault.format.m3/src/main/java/nexusvault/format/m3/v100/InMemoryModelMesh.java
@@ -60,7 +60,7 @@ final class InMemoryModelMesh implements ModelMesh {
 		if ((idx < 0) || (getVertexCount() <= idx)) {
 			throw new IndexOutOfBoundsException(String.format("Idx out of range. Allowed range is [0,%d)", getVertexCount()));
 		}
-		return model.getVertex(structMesh.startVertex, idx);
+		return model.getVertex(structMesh.startVertex + idx);
 	}
 
 	@Override